requirements.txt
streamlit>=1.22.0,<2.0.0
ultralytics>=8.0.124,<9.0.0
Pillow>=10.0.0,<11.0.0
torch>=1.7.0,<3.0.0
opencv-python-headless>=4.1.2,<5.0.0
pandas>=1.3.0,<2.0.0